Move all configuration classes to configuration-package
