+++ service/src/main/java/fi/poltsi/vempain/file/configuration/AsyncConfig.java
@@ -1,4 +1,4 @@
-package fi.poltsi.vempain.file.config;
+package fi.poltsi.vempain.file.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableAsync;
