+++ service/src/main/java/fi/poltsi/vempain/file/configuration/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package fi.poltsi.vempain.file;
+package fi.poltsi.vempain.file.configuration;
 
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
 import io.swagger.v3.oas.annotations.security.SecurityScheme;
